tighten `InsertQueryBuilder.values(...)`.
