+++ test/node/src/error-stack.test.ts
@@ -30,7 +30,7 @@ for (const dialect of BUILT_IN_DIALECTS) {
         ctx.db.deleteFrom('person').where(sql`doesnt_exists`, '=', 1),
         ctx.db
           .insertInto('person')
-          .values({ first_name: sql`doesnt_exists`, gender: 'other' }),
+          .values({ first_name: sql<string>`doesnt_exists`, gender: 'other' }),
 
         {
           execute: () => sql`select doesnt_exists`.execute(ctx.db),
