+++ test/node/src/replace.test.ts
@@ -72,7 +72,7 @@ describe(`${dialect}: replace`, () => {
       first_name: ctx.db
         .selectFrom('pet')
         .select(sql<string>`max(name)`.as('max_name')),
-      last_name: sql`concat('Bar', 'son')`,
+      last_name: sql<string>`concat('Bar', 'son')`,
       gender: 'other',
     })
